refactor: move data files
