+++ src/components/home/slideOurTeam.tsx
@@ -3,7 +3,7 @@ import '../../styles/home/slideOurTeam.scss';
 
 import {TeamCard} from './teamCard';
 import {TeamCarousel} from './teamCarousel';
-import {fondTeam} from './fondTeam';
+import {fondTeam} from '@/fondTeam';
 
 const OPTIONS: EmblaOptionsType = {};
